Add proper seperation between Pickup/Delivery orders, cleaned up UI

+++ checkout.php
@@ -24,7 +24,7 @@ $getOrderID = mysqli_query($dbb, "SELECT MAX(order_id) AS order_id_max FROM orde
 $order_id_max = $getOrderID->fetch_array()['order_id_max'];
 $order_id = $order_id_max + 1;
 
-if (isset($_GET['checkOut'])) {
+if (isset($_GET['checkOutPickup'])) {
   if (isset($_COOKIE['id'])) {
     $query = $db->query('SELECT id, name, username, address FROM users WHERE id = "' . $_COOKIE['id'] . '"');
     $result = $query->fetchAll();
@@ -34,18 +34,67 @@ if (isset($_GET['checkOut'])) {
     }
     $sql = "SELECT * FROM shopping_cart WHERE user_id =" . $_COOKIE['id'] . "";
     $result = $conn->query($sql);
+    $orderTotal = 0;
+    $orderQty = 0;
+
     if ($result->num_rows > 0) {
       for ($i = 0; $i < mysqli_num_rows($result); $i++) {
         $row = mysqli_fetch_assoc($result);
         $product_id = $row['product_id'];
         $product_name = $row['product_name'];
         $product_price = $row['product_price'];
+        $orderTotal += $product_price;
         $product_quantity = $row['product_quantity'];
-
-        $addToList = "INSERT INTO orders_list(product_id,product_name,product_price,product_quantity,order_id,user_id,address) 
-        VALUES ('$product_id','$product_name','$product_price','$product_quantity','$order_id','$user_id','$address')";
-        $conn->query($addToList);
+        $orderQty += $product_quantity;
+       
       }
+      $date = date('d-m-y h:i:s');
+      $addToList = "INSERT INTO orders_list(product_price,product_quantity,order_id,user_id,address,order_date,order_type) 
+      VALUES ('$orderTotal','$orderQty','$order_id','$user_id','$address','$date','Self-Pickup')";
+      $conn->query($addToList);
+      $checkOutSuccessful = true;
+      if ($checkOutSuccessful == true) {
+        $deleteCart = "DELETE FROM shopping_cart WHERE user_id =" . $_COOKIE['id'] . "";
+        $conn->query($deleteCart);
+        echo "<script type='text/javascript'>
+        $(document).ready(function(){
+        $('#CheckoutSuccess').modal('show');
+        });
+        </script>";
+      }
+    }
+  } else {
+  }
+}
+if (isset($_GET['checkOutDelivery'])) {
+  if (isset($_COOKIE['id'])) {
+    $query = $db->query('SELECT id, name, username, address FROM users WHERE id = "' . $_COOKIE['id'] . '"');
+    $result = $query->fetchAll();
+    foreach ($result as $index => $user) {
+      $user_id = $user['id'];
+      $address = $user['address'];
+    }
+    $sql = "SELECT * FROM shopping_cart WHERE user_id =" . $_COOKIE['id'] . "";
+    $result = $conn->query($sql);
+    $orderTotal = 0;
+    $orderQty = 0;
+
+    if ($result->num_rows > 0) {
+      for ($i = 0; $i < mysqli_num_rows($result); $i++) {
+        $row = mysqli_fetch_assoc($result);
+        $product_id = $row['product_id'];
+        $product_name = $row['product_name'];
+        $product_price = $row['product_price'];
+        $orderTotal += $product_price;
+        $product_quantity = $row['product_quantity'];
+        $orderQty += $product_quantity;
+       
+      }
+      $date = date('d-m-y h:i:s');
+      $orderTotal += 5;
+      $addToList = "INSERT INTO orders_list(product_price,product_quantity,order_id,user_id,address,order_date,order_type) 
+      VALUES ('$orderTotal','$orderQty','$order_id','$user_id','$address','$date','Delivery')";
+      $conn->query($addToList);
       $checkOutSuccessful = true;
       if ($checkOutSuccessful == true) {
         $deleteCart = "DELETE FROM shopping_cart WHERE user_id =" . $_COOKIE['id'] . "";
